# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OurFamRoots API
  version: "1.0.0"
  description: |
    ## OurFamRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
    personId:
      name: personId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Person identifier
...
  schemas:
...
    NameType:
      type: string
      enum: [BIRTH, MARRIED, ALSO_KNOWN_AS, NICKNAME, RELIGIOUS, ADOPTED, OTHER]
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://ourfamroots.com/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    PersonName:
      type: object
      required: [id, nameType, isPrimary]
      properties:
        id:
          type: string
          format: uuid
        nameType:
          $ref: "#/components/schemas/NameType"
        givenName:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        prefix:
          type: string
          nullable: true
        suffix:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        fullNameText:
          type: string
          nullable: true
          description: Free-form override for non-Western names
        isPrimary:
          type: boolean
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true
...
    CreatePersonNameInput:
      type: object
      required: [nameType, isPrimary]
      properties:
        nameType:
          $ref: "#/components/schemas/NameType"
        givenName:
          type: string
          maxLength: 200
          nullable: true
        surname:
          type: string
          maxLength: 200
          nullable: true
        prefix:
          type: string
          maxLength: 50
          nullable: true
        suffix:
          type: string
          maxLength: 50
          nullable: true
        nickname:
          type: string
          maxLength: 200
          nullable: true
        fullNameText:
          type: string
          maxLength: 500
          nullable: true
        isPrimary:
          type: boolean
          default: false
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true
...
    NamespaceSummary:
      type: object
      description: >
        Lightweight namespace (tenant) reference attached to user-facing DTOs
        (e.g. `AdminUser.namespace`, `ActivityItem.namespace`).
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        is_global:
          type: boolean
          default: false
          description: True for the single namespace new users land in by default.
    Namespace:
      type: object
      description: A namespace (tenant). Product-facing name for the internal `tenants` table.
      required: [id, name, slug, is_active, is_global, user_count, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 255 }
        slug: { type: string, minLength: 1, maxLength: 100 }
        is_active: { type: boolean }
        is_global:
          type: boolean
          description: True for the single namespace new users land in by default; cannot be deactivated.
        user_count: { type: integer }
        user_preview:
          type: array
          items: { type: string }
          description: First 3 users' display names in this namespace, for an inline "who's in here" preview.
        created_at: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [id, email, app_role, email_verified, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        given_name: { type: string, nullable: true }
        family_name: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        app_role: { type: string, enum: [SUPER_ADMIN, ADMIN, STANDARD, AUDITOR] }
        email_verified: { type: boolean }
        is_active: { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        namespace:
          allOf:
            - $ref: "#/components/schemas/NamespaceSummary"
          nullable: true
          description: The user's namespace. Populated whenever the caller can see across namespaces (Super Admin).

  # ──────────────────────────────────────────────────────────
  # COMMON RESPONSES
  # ──────────────────────────────────────────────────────────
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /trees/{treeId}/persons/{personId}/names:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: listPersonNames
      summary: List all names for a person
      responses:
        "200":
          description: Person names
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PersonName"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    post:
      tags: [Persons]
      operationId: addPersonName
      summary: Add a name to a person
      description: >
        Requires CONTRIBUTOR or higher role.
        If isPrimary=true, the current primary name is demoted automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonNameInput"
      responses:
        "201":
          description: Name added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonName"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /trees/{treeId}/persons/{personId}/names/{nameId}:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
      - name: nameId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Persons]
      operationId: updatePersonName
      summary: Update a person's name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonNameInput"
      responses:
        "200":
          description: Name updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonName"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    delete:
      tags: [Persons]
      operationId: deletePersonName
      summary: Delete a name record
      description: Cannot delete the only primary name
      responses:
        "204":
          description: Name deleted
        "400":
          description: Cannot delete the only primary name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...
  /admin/namespaces:
    get:
      tags: [Admin, Namespaces]
      operationId: listNamespaces
      summary: List namespaces, paginated and searchable (Super Admin only)
      description: Requires `SUPER_ADMIN` app role.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
          description: Search by namespace name or slug (case-insensitive)
      responses:
        "200":
          description: Paginated namespace list
          content:
            application/json:
              schema:
                type: object
                required: [total, items, page, page_size, total_pages]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Namespace"
                  page: { type: integer }
                  page_size: { type: integer }
                  total_pages: { type: integer }
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"

    post:
      tags: [Admin, Namespaces]
      operationId: createNamespace
      summary: Create a new namespace (Super Admin only)
      description: |
        Requires `SUPER_ADMIN` app role. `slug` must be unique across all
        namespaces and match `^[a-z0-9]+(-[a-z0-9]+)*$`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                slug:
                  type: string
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
      responses:
        "201":
          description: Namespace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Namespace"
        "409":
          description: A namespace with this slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /admin/namespaces/{namespaceId}:
    parameters:
      - name: namespaceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Admin, Namespaces]
      operationId: updateNamespace
      summary: Rename or activate/deactivate a namespace (Super Admin only)
      description: |
        Requires `SUPER_ADMIN` app role. All body fields are optional.

        **Restriction:** the Global namespace (the single namespace new users
        land in by default) cannot be deactivated → `400`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                is_active: { type: boolean }
      responses:
        "200":
          description: Updated namespace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Namespace"
        "400":
          description: The Global namespace cannot be deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /admin/namespaces/{namespaceId}/users:
    parameters:
      - name: namespaceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Admin, Namespaces]
      operationId: listNamespaceUsers
      summary: List users in any namespace (Super Admin only)
      description: Requires `SUPER_ADMIN` app role.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
          description: Search by name or email (case-insensitive)
      responses:
        "200":
          description: Paginated user list for this namespace
          content:
            application/json:
              schema:
                type: object
                required: [total, items, page, page_size, total_pages]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  page: { type: integer }
                  page_size: { type: integer }
                  total_pages: { type: integer }
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...